Move tests from base_phone to crm_phone (to benefit from crm_phone_validation)
