--- base_phone/tests/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import test_phone
